Split grids into their own files
