+++ src/Barbs.ts
@@ -7,7 +7,8 @@ import { MapLikeType } from "./Map";
 import { BillboardSpec, RenderMethodArg, TypedArray, WebGLAnyRenderingContext } from "./AutumnTypes";
 import { Color } from "./Color";
 import { ColorMap } from "./Colormap";
-import { AutoZoomGrid, Grid } from "./Grid";
+import { Grid } from "./grids/Grid";
+import { AutoZoomGrid } from "./grids/AutoZoom";
 
 const BASE_BARB_DIMS: BillboardSpec = {
     BB_WIDTH: 85,
