+++ src/StationPlot.ts
@@ -3,7 +3,8 @@ import { RenderMethodArg, WebGLAnyRenderingContext } from "./AutumnTypes";
 import { MapLikeType } from "./Map";
 import { PlotComponent } from "./PlotComponent";
 import { normalizeOptions } from "./utils";
-import { AutoZoomGrid, Grid } from "./Grid";
+import { Grid } from "./grids/Grid";
+import { AutoZoomGrid } from "./grids/AutoZoom";
 import { RawObsField } from "./RawField";
 import { HorizontalAlign, TextCollection, TextCollectionOptions, TextSpec, VerticalAlign } from "./TextCollection";
 import { Float16Array } from "@petamoriken/float16";
