+++ src/index.ts
@@ -15,7 +15,12 @@ import { Color } from "./Color";
 import { makeColorBar, makePaintballKey, ColorbarOrientation, ColorbarTickDirection, ColorBarOptions, PaintballKeyOptions } from "./ColorBar";
 import { LineStyle } from "./PolylineCollection";
 import { RawScalarField, RawVectorField, RawProfileField, VectorRelativeTo, RawVectorFieldOptions, RawObsField, ObsRawData} from "./RawField";
-import { Grid, GridType, StructuredGrid, PlateCarreeGrid, PlateCarreeRotatedGrid, LambertGrid, UnstructuredGrid } from './Grid'
+import { Grid, GridType } from "./grids/Grid";
+import { StructuredGrid } from "./grids/StructuredGrid";
+import { PlateCarreeGrid } from "./grids/PlateCarreeGrid";
+import { PlateCarreeRotatedGrid } from "./grids/PlateCarreeRotatedGrid";
+import { LambertGrid } from "./grids/LambertGrid";
+import { UnstructuredGrid } from "./grids/UnstructuredGrid";
 
 import { initMSModule, FieldContourOpts } from './ContourCreator';
